+++ NEWS.md
@@ -1,6 +1,6 @@
 # teal.code 0.3.0.9008
 
-* `eval_code` doesn't render plots in a graphical device.  
+* `dev_suppress` has been added to suppress rendering of plots on IDE.    
 * `chunks` have been removed. The new `qenv` object should be used instead. See the new `qenv` vignette in the package for further details.
 
 # teal.code 0.3.0